Default `:: replaceStrings()`
