+++ src/Content/MemoryContentStorageHandler.php
@@ -5,7 +5,6 @@ namespace Kirby\Content;
 use Kirby\Cache\MemoryCache;
 use Kirby\Cms\Language;
 use Kirby\Cms\ModelWithContent;
-use Kirby\Toolkit\A;
 
 /**
  * @package   Kirby Content
@@ -78,28 +77,6 @@ class MemoryContentStorageHandler extends ContentStorageHandler
 	{
 		$this->ensure($versionId, $language);
 		return $this->cache->get($this->cacheId($versionId, $language));
-	}
-
-	/**
-	 * Searches and replaces one or multiple strings
-	 *
-	 * @throws \Kirby\Exception\NotFoundException If the version does not exist
-	 */
-	public function replaceStrings(
-		VersionId $versionId,
-		Language $language,
-		array $map
-	): void {
-		$fields = $this->read($versionId, $language);
-		$fields = A::map(
-			$fields,
-			fn ($field) => str_replace(
-				array_keys($map),
-				array_values($map),
-				$field
-			)
-		);
-		$this->update($versionId, $language, $fields);
 	}
 
 	/**
